Move files to packages
Files too cluttered

Hard to view

Split into various packages for
easier viewing
