--- src/main/java/DukeException.java
@@ -1,9 +1,0 @@
-/**
- * The DukeException Class extends Exception and is
- * used to denote any exception specific to Duke
- */
-public class DukeException extends Exception{
-    public DukeException(String msg) {
-        super(msg);
-    }
-}
