--- src/main/java/UnknownCommandException.java
@@ -1,9 +1,0 @@
-/**
- * The UnknownCommandException extends DukeException and is used
- * to denote that Duke does not understand that particular command
- */
-public class UnknownCommandException extends DukeException{
-    public UnknownCommandException() {
-        super("I do not understand this command");
-    }
-}
