+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+import duke.task.Task;
+
 import java.io.IOException;
 
 public class AddCommand extends Command{
