+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+
 public class ExitCommand extends Command {
 
     public ExitCommand() {
