+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+
 public class UnknownCommand extends Command {
 
     public UnknownCommand() {
