+++ src/main/java/duke/exception/EmptyDescriptionException.java
@@ -1,5 +1,7 @@
+package duke.exception;
+
 /**
- * The EmptyDescriptionException Class extends DukeException and
+ * The duke.exception.EmptyDescriptionException Class extends duke.exception.DukeException and
  * is used to denote when a command has a missing description
  */
 public class EmptyDescriptionException extends DukeException{
