+++ src/main/java/duke/exception/MissingIndexException.java
@@ -1,5 +1,7 @@
+package duke.exception;
+
 /**
- * The MissingIndexException Class extends DukeException and is
+ * The duke.exception.MissingIndexException Class extends duke.exception.DukeException and is
  * used to denote that there is a missing parameter or index
  */
 public class MissingIndexException extends DukeException{
