+++ src/main/java/duke/task/Event.java
@@ -1,5 +1,7 @@
+package duke.task;
+
 /**
- * The Event class extends Task. An event has 2 extra fields
+ * The duke.task.Event class extends duke.task.Task. An event has 2 extra fields
  * of start and end time
  *
  * @author Zi Xiang
@@ -9,7 +11,7 @@ public class Event extends Task{
     protected String start;
     protected String end;
 
-    /** Constructor for Event */
+    /** Constructor for duke.task.Event */
     public Event(String done, String description, String start, String end) {
         super(description, done);
         this.start = start;
