+++ src/main/java/duke/task/Todo.java
@@ -1,12 +1,14 @@
+package duke.task;
+
 /**
- * The T*odo class extends Task
+ * The T*odo class extends duke.task.Task
  *
  * @author Zi Xiang
  * @version CS2103 AY23/24 Sem 1
  */
 public class Todo extends Task{
 
-    /** Constructor for Todo */
+    /** Constructor for duke.task.Todo */
     public Todo(String done, String description) {
         super(description, done);
     }
